my-container-infrastructure/requirements.txt
aws-cdk-lib==2.44.0
constructs>=10.0.0,<11.0.0
cdk-monitoring-constructs>=1.0.0,<2.0.0
